Development

+++ src/lib/Model/AdditionalEarningsInputModel.php
@@ -76,7 +76,7 @@ class AdditionalEarningsInputModel implements ModelInterface, ArrayAccess
       * @var string[]
       */
     protected static $swaggerFormats = [
-        'id' => 'int32',
+        'id' => 'int64',
         'pay_category_id' => 'int32',
         'employee_id' => 'int32',
         'amount' => 'double',
